adding AppSidebar component and MailPage, adding AppSidebar to every page in App.tsx, moving app components
