+++ frontend/src/App.tsx
@@ -4,21 +4,23 @@ import { BrowserRouter as Router, Switch, Route } from "react-router-dom"
 import TokenPage from "./FullPageRoutes/TokenPage"
 import ProfilePage from "./FullPageRoutes/Profile"
 import MatchingPage from "./FullPageRoutes/Matching"
-import { Box, makeStyles } from "@material-ui/core"
+import { Box, makeStyles, Grid } from "@material-ui/core"
 import RecordPage from "./FullPageRoutes/RecordPage"
 import ProvideContext from "./components/utils/ProvideContext"
-import AppHeader from "./components/AppHeader"
-import AppFooter from "./components/AppFooter"
+import AppHeader from "./components/app/AppHeader"
+import AppFooter from "./components/app/AppFooter"
 import { ProtectedRoute } from "./components/ProtectedRoute"
-import Dashboard from "./FullPageRoutes/Dashboard"
+import AppSideBar from "./components/app/AppSidebar"
+import MailPage from "./FullPageRoutes/MailPage"
 import VideoViewPage from "./FullPageRoutes/VideoViewPage"
 
 const useStyles = makeStyles({
   app: {
-    height: "40rem",
+    minHeight: "38rem",
     display: "flex",
     flexDirection: "column",
-    paddingTop: "50px",
+    padding: "3rem 0",
+    backgroundColor: "#ede8db",
   },
   appRoot: {
     minHeight: "55rem",
@@ -27,10 +29,41 @@ const useStyles = makeStyles({
     minHeight: "49rem",
     width: "100%",
   },
+  sidebar: {
+    borderRight: "1px solid white",
+  },
 })
+
+type Page = {
+  path: string
+  pageJsx: JSX.Element
+}
 
 export default function App(): JSX.Element {
   const classes = useStyles()
+
+  const pageList: Page[] = [
+    {
+      path: "/record/:receiver",
+      pageJsx: <RecordPage />,
+    },
+    {
+      path: "/matches/",
+      pageJsx: <MatchingPage />,
+    },
+    {
+      path: "/profile/",
+      pageJsx: <ProfilePage />,
+    },
+    {
+      path: "/watch/",
+      pageJsx: <VideoViewPage />,
+    },
+    {
+      path: "/",
+      pageJsx: <MailPage />,
+    },
+  ]
 
   return (
     <ProvideContext>
@@ -41,21 +74,18 @@ export default function App(): JSX.Element {
             <Route path="/tokens">
               <TokenPage />
             </Route>
-            <ProtectedRoute path="/record/:receiver">
-              <RecordPage />
-            </ProtectedRoute>
-            <ProtectedRoute path="/matches">
-              <MatchingPage />
-            </ProtectedRoute>
-            <ProtectedRoute path="/profile">
-              <ProfilePage />
-            </ProtectedRoute>
-            <ProtectedRoute path="/watch">
-              <VideoViewPage />
-            </ProtectedRoute>
-            <ProtectedRoute path="/">
-              <Dashboard />
-            </ProtectedRoute>
+            {pageList.map(({ path, pageJsx }: Page, index: number) => (
+              <ProtectedRoute key={`page-${index}`} path={path}>
+                <Grid container direction="row">
+                  <Grid item className={classes.sidebar} xs={2}>
+                    <AppSideBar />
+                  </Grid>
+                  <Grid item xs={10}>
+                    {pageJsx}
+                  </Grid>
+                </Grid>
+              </ProtectedRoute>
+            ))}
           </Switch>
         </Box>
         <AppFooter />
