+++ frontend/src/components/app/AppFooter.tsx
